SoundBlaster: Remove floating point usage

+++ src/device/sound/speaker/PcSpeaker.h
@@ -65,45 +65,45 @@ public:
      */
     static void off();
 
-    static const constexpr float C0 = 130.81;
-    static const constexpr float C0X = 138.59;
-    static const constexpr float D0 = 146.83;
-    static const constexpr float D0X = 155.56;
-    static const constexpr float E0 = 164.81;
-    static const constexpr float F0 = 174.61;
-    static const constexpr float F0X = 185.00;
-    static const constexpr float G0 = 196.00;
-    static const constexpr float G0X = 207.65;
-    static const constexpr float A0 = 220.00;
-    static const constexpr float A0X  = 233.08;
-    static const constexpr float B0 = 246.94;
+    static const constexpr uint32_t C0 = 130.81;
+    static const constexpr uint32_t C0X = 138.59;
+    static const constexpr uint32_t D0 = 146.83;
+    static const constexpr uint32_t D0X = 155.56;
+    static const constexpr uint32_t E0 = 164.81;
+    static const constexpr uint32_t F0 = 174.61;
+    static const constexpr uint32_t F0X = 185.00;
+    static const constexpr uint32_t G0 = 196.00;
+    static const constexpr uint32_t G0X = 207.65;
+    static const constexpr uint32_t A0 = 220.00;
+    static const constexpr uint32_t A0X  = 233.08;
+    static const constexpr uint32_t B0 = 246.94;
 
-    static const constexpr float C1 = 261.63;
-    static const constexpr float C1X = 277.18;
-    static const constexpr float D1 = 293.66;
-    static const constexpr float D1X = 311.13;
-    static const constexpr float E1 = 329.63;
-    static const constexpr float F1 = 349.23;
-    static const constexpr float F1X = 369.99;
-    static const constexpr float G1 = 391.00;
-    static const constexpr float G1X = 415.30;
-    static const constexpr float A1 = 440.00;
-    static const constexpr float A1X = 466.16;
-    static const constexpr float B1 = 493.88;
+    static const constexpr uint32_t C1 = 261.63;
+    static const constexpr uint32_t C1X = 277.18;
+    static const constexpr uint32_t D1 = 293.66;
+    static const constexpr uint32_t D1X = 311.13;
+    static const constexpr uint32_t E1 = 329.63;
+    static const constexpr uint32_t F1 = 349.23;
+    static const constexpr uint32_t F1X = 369.99;
+    static const constexpr uint32_t G1 = 391.00;
+    static const constexpr uint32_t G1X = 415.30;
+    static const constexpr uint32_t A1 = 440.00;
+    static const constexpr uint32_t A1X = 466.16;
+    static const constexpr uint32_t B1 = 493.88;
 
-    static const constexpr float C2 = 523.25;
-    static const constexpr float C2X = 554.37;
-    static const constexpr float D2 = 587.33;
-    static const constexpr float D2X = 622.25;
-    static const constexpr float E2 = 659.26;
-    static const constexpr float F2 = 698.46;
-    static const constexpr float F2X = 739.99;
-    static const constexpr float G2 = 783.99;
-    static const constexpr float G2X = 830.61;
-    static const constexpr float A2 = 880.00;
-    static const constexpr float A2X = 923.33;
-    static const constexpr float B2 = 987.77;
-    static const constexpr float C3 = 1046.50;
+    static const constexpr uint32_t C2 = 523.25;
+    static const constexpr uint32_t C2X = 554.37;
+    static const constexpr uint32_t D2 = 587.33;
+    static const constexpr uint32_t D2X = 622.25;
+    static const constexpr uint32_t E2 = 659.26;
+    static const constexpr uint32_t F2 = 698.46;
+    static const constexpr uint32_t F2X = 739.99;
+    static const constexpr uint32_t G2 = 783.99;
+    static const constexpr uint32_t G2X = 830.61;
+    static const constexpr uint32_t A2 = 880.00;
+    static const constexpr uint32_t A2X = 923.33;
+    static const constexpr uint32_t B2 = 987.77;
+    static const constexpr uint32_t C3 = 1046.50;
 
 private:
